Create asdf.py
This is an example program that can be compiled into a pyd file. The pyd file can be imported when running in Windows cmd. Online sources say it also works in Jupyter Notebook. However, it didn't pass Spyer in my tests, potentially because of different Python compiler version or something.